Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -6,8 +6,8 @@ char* welcome() {
 
 int main() {
    // printf("%s", welcome());
-   int a,b;
-   scanf("%f %f",&a,&b);
+   float a,b;
+   scanf("%.2f %.2f",&a,&b);
    printf("Product: %.2f",a*b);
     return 0;
 }
